src-tauri/requirements.txt
google-auth-oauthlib==1.2.4
google-api-python-client==2.188.0
google-auth==2.47.0
requests-oauthlib==2.0.0
